Join functional and integration tests of BlockHeadersReader
